time int->int64

+++ api.go
@@ -33,7 +33,7 @@ type Story struct {
 	Parts       []int  `json:"parts"`
 	Score       int    `json:"score"`
 	Text        string `json:"text"`
-	Time        int    `json:"time"`
+	Time        int64  `json:"time"`
 	Title       string `json:"title"`
 	Type        string `json:"type"`
 	URL         string `json:"url"`
